+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 - Improved folder structure, easier to maintain.
 - bette
 - Retention limits to be little safer.
+- Data view
 
 == Fixes ==
 - Fixed bnet chat export/copy window.
 - Combat handling issues.
+- Being unable to view censored messaged.